// src/api/models/RecommendationTabDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecommendationTabDto = {
    id: string;
    categoryKey?: Record<string, any>;
    label: string;
    sort: number;
    enabled: boolean;
    icon?: string;
    createdAt: string;
    updatedAt: string;
};